Move get_repositories to github_knowledge
