+++ srcopsmetrics/github_knowledge.py
@@ -73,6 +73,28 @@ class GitHubKnowledge:
         return repo
 
     @staticmethod
+    def get_repositories(repository: str = None, organization: str = None) -> List[str]:
+        """Get overall repositories to be inspected.
+
+        :param repository:str:
+        :param organization:str:
+
+        :rtype: List of all repositories (repository + repositories in organization)
+        """
+        repos = []
+        gh = Github(login_or_token=_GITHUB_ACCESS_TOKEN, timeout=50)
+
+        if repository is not None:
+            repos.append(gh.get_repo(repository).full_name)
+
+        if organization is not None:
+            for r in gh.get_organization(organization).get_repos():
+                repos.append(r.full_name)
+
+        _LOGGER.info("Overall repositories found: %d" % len(repos))
+        return repos
+
+    @staticmethod
     def assign_pull_request_size(lines_changes: int) -> str:
         """Assign size of PR is label is not provided."""
         if lines_changes > 1000:
